+++ Templates/src/Core/PainKiller.PowerCommands.Configuration/ENUMS/HideToollbarOption.cs
@@ -1,8 +1,0 @@
-﻿namespace $safeprojectname$.Enums;
-
-public enum HideToollbarOption
-{
-    Never,
-    OnTextChange,
-    OnCommandHighlighted
-}