+++ Templates/src/PainKiller.PowerCommands.Bootstrap/Startup.cs
@@ -1,4 +1,4 @@
-﻿using Microsoft.Extensions.Logging;
+using Microsoft.Extensions.Logging;
 using PainKiller.PowerCommands.Configuration;
 using PainKiller.PowerCommands.Configuration.DomainObjects;
 using PainKiller.PowerCommands.Core.Managers;