+++ Templates/src/PainKiller.PowerCommands.MyExampleCommands/COMMANDS/ListCommand.cs
@@ -8,10 +8,10 @@ public class ListCommand : CommandBase<PowerCommandsConfiguration>
 
     public override RunResult Run()
     {
-        var selectedItems = DialogService.ListDialog("Which teams participated in the Stanley Cup final season 2022/23?", new() { "New York Rangers", "Colorado Avalanche", "Vegas Knights", "Pittsburgh Penguins", "Florida Panthers", "Detroit Red Wings", "Toronto Maple Leafs" });
+        var selectedItems = DialogService.ListDialog("Which teams participated in the Stanley Cup final season 2022/23?", new() { "New York Rangers", "Colorado Avalanche", "Vegas Knights", "Pittsburgh Penguins", "Florida Panthers", "Detroit Red Wings", "Toronto Maple Leafs" }, multiSelect: true);
         WriteHeadLine("You selected");
-        foreach (var item in selectedItems) WriteLine(item);
-        if (selectedItems.Count == 2 && selectedItems.Any(t => t == "Vegas Knights") && selectedItems.Any(t => t == "Florida Panthers"))
+        foreach (var item in selectedItems) WriteLine(item.Value);
+        if (selectedItems.Count == 2 && selectedItems.Any(t => t.Value == "Vegas Knights") && selectedItems.Any(t => t.Value == "Florida Panthers"))
         {
             WriteSuccessLine("\nYour absolutely right, and Vegas Knights won the Stanley Cup title!");
         }