+++ Templates/src/PainKiller.PowerCommands.MyExampleCommands/PowerCommandServices.cs
@@ -21,7 +21,6 @@ public class PowerCommandServices : IExtendedPowerCommandServices<PowerCommandsC
         DefaultConsoleService = ConsoleService.Service;
         
         var suggestions = new List<string>(Runtime.CommandIDs);
-        suggestions.AddRange(Runtime.CommandIDs.Select(s => $"describe {s}").ToList());
         suggestions.AddRange(Runtime.Commands.Where(c => !string.IsNullOrEmpty(c.GetDefaultParameter())).Select(c => $"{c.Identifier} {c.GetDefaultParameter()}").ToList());
 
         ReadLineService.InitializeAutoComplete(history: new string[]{},suggestions: suggestions.ToArray());