+++ src/PainKiller.PowerCommands/Core/PainKiller.PowerCommands.Security/Managers/ChecksumManager.cs
@@ -4,9 +4,9 @@ using System.Text;
 
 namespace PainKiller.PowerCommands.Security.Managers;
 
-internal static class ChecksumManager
+public static class ChecksumManager
 {
-    internal static string CalculateMd5ForFile(string filename)
+    public static string CalculateMd5ForFile(string filename)
     {
         if (!File.Exists(filename)) return "file missing...";
         using var md5 = MD5.Create();
@@ -14,7 +14,7 @@ internal static class ChecksumManager
         var hash = md5.ComputeHash(stream);
         return BitConverter.ToString(hash).Replace("-", "").ToLowerInvariant();
     }
-    internal static string CalculateMd5<TItem>(TItem content)
+    public static string CalculateMd5<TItem>(TItem content)
     {
         var options = new JsonSerializerOptions { WriteIndented = true, IncludeFields = true };
         var jsonString = JsonSerializer.Serialize(content, options);