+++ src/PainKiller.PowerCommands/PainKiller.PowerCommands.Bootstrap/PowerCommandsManager.partial.cs
@@ -1,16 +1,17 @@
+using PainKiller.PowerCommands.Core.Managers;
 using PainKiller.PowerCommands.Core.Services;
 
 namespace PainKiller.PowerCommands.Bootstrap;
 public partial class PowerCommandsManager
 {
     //private static bool _hasRunOnce = false;
-    private void RunCustomCode()
+    private void RunCustomCode(RunFlowManager runFlow)
     {
         //Add your custom code here, som example code below shows an example on how could find a command an run a method, could be a startup toolbar for example.
         //This example is using the configuration startupToolbar in the PowerCommandsConfiguration.yaml file.
         //if (!_hasRunOnce)
         //{
-        //    DialogService.DrawToolbar(Services.ExtendedConfiguration.StartupToolbar);
+        //    ToolbarService.DrawToolbar(Services.ExtendedConfiguration.StartupToolbar);
         //    _hasRunOnce = true;
         //}
     }